Combining test files and clearing instance in tearDown
